% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2mat.R
\name{bed2mat}
\alias{bed2mat}
\title{Subset a BED file and convert to SNP matrix}
\usage{
bed2mat(bed, genos = NULL, min_maf = 0, max_het = 1)
}
\arguments{
\item{bed}{A bed.matrix object created with package 'gaston'}

\item{genos}{A character vector listing genotypes (i.e. individuals) to
include in the final genotypic matrix}

\item{min_maf}{Number between 0 and 0.5 - the minimum minor-allele-frequency
threshold to apply after subsetting}

\item{max_het}{Number between 0 and 1 - the maximum SNPwise heterozygosity to
allow after subsetting}
}
\value{
A subsetted, possibly filtered SNP matrix created from the input
  BED file
}
\description{
Subset a BED file and convert to SNP matrix
}
\details{
This function takes a BED (PLINK binary pedigree/map) file,
  optionally subsets individuals from it, and optionally filters SNPs by minor-
  allele frequency and/or percent heterozygous calls. It then outputs a SNP 
  matrix. It is advised to re-filter SNPs whenever subsetting individuals, 
  as allele frequencies and the proportion of heterozygous calls may change.
}
