% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coindex.R
\name{coindex}
\alias{coindex}
\title{Calculate Coincidence Index}
\usage{
coindex(obs, pred, si, best = "high")
}
\arguments{
\item{obs}{A named vector of observed phenotypic data}

\item{pred}{A named vector of predicted phenotypic data}

\item{si}{Selection index - value from 0 to 1}

\item{best}{String or number signifying what phenotypic values are considered
 "best." Must be one of the following:
\enumerate{
  \item "high" to perform directional selection of genotypes with highest trait values
  \item "low" to perform directional selection of genotypes with lowest trait values
  \item "mean" to perform stabilizing selection around the mean values of the observed and predicted data
  \item "median" to perform stabilizing selection around the median values of the observed and predicted data
  \item numeric, in which case it is assumed that the ideal phenotypic value has been supplied
}}
}
\value{
The calculated coincidence index
}
\description{
Calculate Coincidence Index
}
\details{
This function calculates the coincidence index as defined by Hamblin
  and Zimmerman, 1986. (https://doi.org/10.1002/9781118061015.ch8). In the context
  of this function, its intended use is to compare direct phenotypic selection
  with indirect selection based upon genomic data. However, it can be used to
  compare selections made using any two different methods. The formula for the
  coincidence index is
  \deqn{CI = (C - R) / (T - R)}
  where T is the number of observed genotypes selected using the selection index,
  R is the expected number of genotypes among these correctly selected due to 
  chance, and C is the number of coincident genotypes selected using both methods. 
  
  For instance, if we have a population of 100 genotypes and a selection index of
  0.1, T = 100 * 0.1 = 10 and R = 10 * 0.1 = 1. The value of C will depend upon
  how many genotypes are selected based upon both observed and predicted data.
  
  A value of 1 indicates identical selections made using the observed and
  predicted phenotypic values, while a value of 0 indicates that selections
  based on predicted values were no better than random chance. Note that CI values
  can be negative. Small samples sizes will lead to high CI variance.
}
