% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_bar.R
\name{color_bar}
\alias{color_bar}
\title{Create a Color Bar}
\usage{
color_bar(
  pal,
  min,
  max = -min,
  width = 1,
  nticks = 11,
  ticks = seq(min, max, len = nticks),
  title = ""
)
}
\arguments{
\item{pal}{Input color palette}

\item{min}{Minimum value for colorbar axis}

\item{max}{Maximum value for colorbar axis}

\item{width}{Value specifying width of the color bar rectangle - must be
between 0 and 10}

\item{nticks}{Integer specifying number of tick marks to label color bar axis
with}

\item{ticks}{Numeric vector specifying position of tick marks, for instance as
produced by seq()}

\item{title}{Title for the color bar plot}
}
\description{
Create a Color Bar
}
\details{
All credit goes to Stackoverflow user John Colby:
https://stackoverflow.com/questions/9314658/colorbar-from-custom-colorramppalette
}
