% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cullis_H2.R
\name{Cullis_H2}
\alias{Cullis_H2}
\title{Calculate Generalized Heritability from lme4 Model}
\usage{
Cullis_H2(model, geno_name = "GENO")
}
\arguments{
\item{model}{A lme4 model object}

\item{geno_name}{A string denoting the label of the random genotypic effect in the
supplied lme4 model object}
}
\value{
A list containing the following elements:
\itemize{
\item SED A vector containing the standard errors of the genotypic random effects
(the length of this vector is the number of genotypes supplied to the model)
\item H2 The generalized heritability estimate
}
}
\description{
Calculate Generalized Heritability from lme4 Model
}
\details{
This function calculates generalized heritability using the method of
Cullis et al., 2006 (\url{https://doi.org/10.1198/108571106X154443}). This
method can be used in unbalanced applications where the traditional entry
mean heritability calculation will give biased estimates. The method of doing
this using lme4 is detailed at \url{https://shantel-martinez.github.io/resources.html}
}
