% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigstrat.R
\name{eigstrat}
\alias{eigstrat}
\title{Perform Eigenstrat marker PCA}
\usage{
eigstrat(geno, snps = "rows")
}
\arguments{
\item{geno}{A matrix of SNP data encoded in minor-allele dosage format:
2 = homozygous for minor allele
1 = heterozygous
0 = homozygous for major allele
NA = missing data}

\item{snps}{Either "rows" to indicate that SNPs are in rows and individuals 
are in columns, or "cols" to indicate the SNPs are in columns and
individuals are in rows}
}
\value{
The Eigen decomposition of the input genotypic matrix
}
\description{
Perform Eigenstrat marker PCA
}
\details{
This function performs principle component analysis on a matrix of
  SNP data in the method of EIGENSTRAT (Price et al, 2006; doi:10.1038/ng1847)
}
