% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_connect.R
\name{env_connect}
\alias{env_connect}
\title{Evaluate connectivity between environments}
\usage{
env_connect(pheno, env, geno, output = "dataframe")
}
\arguments{
\item{pheno}{Dataframe containing one column for environment identifier, and
one column for genotype identifier.}

\item{env}{String containing the name of the column in pheno containing
environment identifiers.}

\item{geno}{String containing the name of the column in pheno containing
genotype identifiers.}

\item{output}{String containing either "dataframe" to return a three-column
dataframe, or "matrix" to return a sparse matrix.}
}
\value{
Either a dataframe or else a sparse matrix containing the number of
genotypes within each environment, and the number of genotypes shared between
each pair of environments.
}
\description{
Evaluate connectivity between environments
}
\details{
This function takes as input a dataframe containing one column 
containing identifiers for environments, and one column containing identifiers
for genotypes, and returns a structure that lists the number of genotypes
evaluated within each environment, and the number of genotypes shared between
each pair of environments. This information can either be returned as a
dataframe with three columns, or else as a matrix.
}
