% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_thresh.R
\name{fdr_thresh}
\alias{fdr_thresh}
\title{FDR threshold and SNP highlighting}
\usage{
fdr_thresh(gwas_results, id, p, meth = "fdr", thresh = 0.05)
}
\arguments{
\item{gwas_results}{A dataframe containing at least the following columns:
1) SNP identifiers
2) GWAS p-values.  These should NOT be -log transformed}

\item{id}{String specifying the name of the column in gwas_results
containing SNP identifiers}

\item{p}{Character specifying the name of the column in gwas_results
containing GWAS output p-values}

\item{meth}{String specifying the method of p-value adjustment to use - 
see output of help(p.adjust) for details}

\item{thresh}{Number between 0 and 1 - the desired FDR threshold}
}
\value{
list consisting of
  \item{p.thresh}{approximation of the FDR threshold for p-value distribution}
  \item{sig.snps}{vector of names of SNPs passing the FDR threshold}
}
\description{
FDR threshold and SNP highlighting
}
\details{
Adjusted p-values are calculated from a vector of p-values output
  from a GWAS analysis. If any adjusted p-values fall below the specified
  threshold, an approximate threshold for adjusted p-values is returned, with
  a list of the IDs of significant SNPs
}
