% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_qtlmap_geno.R
\name{format_qtlmap_geno}
\alias{format_qtlmap_geno}
\title{Generate files for QTL mapping}
\usage{
format_qtlmap_geno(bed, par_a, par_b, rm_mono = TRUE, rm_het = TRUE,
  rm_miss = TRUE, include_pars = TRUE, out_fmt = "rqtl")
}
\arguments{
\item{bed}{A binary ped (bed) object created with the package 'gaston',
likely either using the functions gaston::read.bed.matrix() or
gaston::read.vcf()}

\item{par_a}{String identifying the column in the SNP matrix containing
data for the first parent.}

\item{par_b}{String identifying the column in the SNP matrix containing 
data for the second parent.}

\item{rm_mono}{Logical indicating whether to remove SNPs that are monomorphic
between the two parents.}

\item{rm_het}{Logical indicating whether to remove SNPs that are heterozygous
in either parent.}

\item{rm_miss}{Logical indicating whether to remove SNPs that are missing
in either parent.}

\item{include_pars}{Logical indicating whether to include the parents in the
output data}

\item{out_fmt}{String indicating the format of the output table - one of
either "icimapping" or "rqtl"}
}
\description{
Generate files for QTL mapping
}
\details{
If "rqtl" output format is selected, a table suitable for use as the
  genotypic portion of the separate, rotated (i.e. "csvsr") R/qtl input
  format is produced.
}
