% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_plot_gen.R
\name{ge_plot_gen}
\alias{ge_plot_gen}
\title{Generate GxE Plot for a trait}
\usage{
ge_plot_gen(trait_df, trait_of_int, min_core = 5)
}
\arguments{
\item{trait_df}{- a dataframe or tibble with at least three columns - 
GENOTYPE, ENV, and one or more columns for trait values}

\item{trait_of_int}{- string matching the column name for the trait of
interest in trait_df}

\item{min_core}{- integer specifying the minimum number of "core" genotypes
(i.e. genotypes tested across all environments) required to produce plot. If
the number of core genotypes falls below this number, NULL will be returned}
}
\value{
a ggplot plot object, or NULL
}
\description{
Generate GxE Plot for a trait
}
\details{
This function generates GxE plots in a manner similar to that of
  Finlay and Wilkinson (1963; https://doi.org/10.1071/AR9630742). Briefly,
  a "core" set of genotypes tested across all environments is identified.
  The environments are then arranged in increasing order by the mean values
  of the core genotypes within each environment, and then a regression is fit
  for each individual core genotype.
}
