% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_pca.R
\name{geno_pca}
\alias{geno_pca}
\title{Perform PCA on marker data}
\usage{
geno_pca(geno)
}
\arguments{
\item{geno}{A matrix of SNP data with SNPs in rows, genotypes in columns.
SNPs should be encoded in minor-allele dosage format, so that:
2 = homozygous for minor allele
1 = heterozygous
0 = homozygous for major allele
NA = missing data}
}
\value{
The Eigen decomposition of the input genotypic matrix
}
\description{
Perform PCA on marker data
}
\details{
This function performs principle component analysis on a matrix of
  SNP data. Note that there are a few transpositions required, as R would
  typically not perform the centering/scaling/PCA in the correct orientation
}
