% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt2num.R
\name{gt2num}
\alias{gt2num}
\title{Convert VCF GT fields to numeric format}
\usage{
gt2num(genomat)
}
\arguments{
\item{genomat}{A matrix of VCF-style genotype calls, with SNPs in rows, and
samples/individuals in columns}
}
\value{
A list containing the following elements:
\itemize{
\item genomat Converted numeric genotype matrix
\item removed_loci Vector of loci removed due to having greater than two alleles
}
}
\description{
Convert VCF GT fields to numeric format
}
\details{
This function will convert a matrix of genotype calls from a VCF file
into numeric format. It will convert homozygous reference allele calls to 0,
heterozygous calls to 1, and homozygous alternate allele calls to 2.
Note that the output IS NOT in minor-allele dosage format. See ?flip_alleles.
Loci with more than two alleles will be removed and reported. An input matrix
of the required format can be generated with vcfR::extract.gt(), with
bcftools query, or manually in R.
}
