% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kasp_consensus.R
\name{kasp_consensus}
\alias{kasp_consensus}
\title{Generate consensus calls for KASP marker data}
\usage{
kasp_consensus(
  kasp,
  geno_col,
  data_cols,
  ambiguous = "strict",
  out_format = "KASP"
)
}
\arguments{
\item{kasp}{A dataframe containing KASP-formatted marker data. This should
consist of at least one column with genotype names, and additional columns
containing "X:X" and "Y:Y" for the two homozygous states, "X:Y" or "Y:X"
for the heterozgous states, and "NO CALL" or "NOCALL" for missing data.}

\item{geno_col}{String identifying which column in kasp contains genotype names}

\item{data_cols}{Character vector identifying which columns in kasp will be
retained and treated as marker data columns}

\item{ambiguous}{String consisting of either "strict" or "lax". If "strict",
marker calls with uncertainty (followed by a question mark) will be converted
to missing data. If "lax", uncertain calls will be converted to their
putative call (e.g. "X:X?" will be converted to "X:X")}

\item{out_format}{String consisting of either "KASP", "VCF", or "numeric". If
"KASP", output will be in same format as the input. IF "VCF", output will
be encoded as "0/0" and "1/1" for the homozgous states, "0/1" for the
heterozygous state, and "./." for missing data. If "numeric", output will be
encoded as -1 and 1 for homozygous states, 0 for the heterozygous state,
and NA for missing data.}
}
\value{
Data frame with consensus calling performed, encoded as specified by
  out_format
}
\description{
Generate consensus calls for KASP marker data
}
\details{
kasp_consensus generates consensus calls for KASP marker data when
  some lines have been genotyped multiple times. It does this using a simple
  algorithm in which markers are first converted to numeric format, and then 
  subsequently summed for each genotype. Note that it cannot differentiate
  between randomly missing data and true null alleles. The input data must be
  reasonably well-formatted. The function can handle a variety of different
  formats (e.g. "x:x", "X:X", "NO CALL", "nocall"), as well as uncertain calls
  (e.g. "X:X?") but other more human-friendly data input will simply be
  changed to missing data, with a warning that data has been converted to NA
  by coercion.
}
