% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_pca.R
\name{marker_pca}
\alias{marker_pca}
\title{Perform PCA on marker data}
\usage{
marker_pca(geno, snps = "rows")
}
\arguments{
\item{geno}{A matrix of SNP data with SNPs in minor-allele dosage format:
2 = homozygous for minor allele
1 = heterozygous
0 = homozygous for major allele
NA = missing data}

\item{snps}{Either "rows" to indicate that SNPs are in rows and individuals 
are in columns, or "cols" to indicate the SNPs are in columns and
individuals are in rows}
}
\value{
The Eigen decomposition of the input genotypic matrix
}
\description{
Perform PCA on marker data
}
\details{
This function performs a "quick and dirty" principle component 
  analysis on the input matrix of SNP data. A genetic relationship matrix
  is calculated first, and then an Eigen decomposition is performed on this
  matrix
}
