% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_headrows.R
\name{randomize_headrows}
\alias{randomize_headrows}
\title{Randomize Headrow Trays}
\usage{
randomize_headrows(ents, nreps, tray_cells = 80, tray_cols = 4,
  checks = list("CHECK", 40), seed = NA)
}
\arguments{
\item{ents}{A vector listing names of entries to include in the experiment}

\item{nreps}{An integer specifying the number of replications 
(i.e. full blocks)}

\item{tray_cells}{An integer specifying the number of cells contained by each
headrow tray}

\item{tray_cols}{An integer specifying the number of columns in a headrow
tray}

\item{checks}{A list with two elements:
\enumerate{
  \item A string or character vector listing the names of check(s) to
  include
  \item An integer specifying how often to repeat the checks  
}}

\item{seed}{An optional integer to seed the random number generator}
}
\value{
A dataframe containing the randomized headrows and all pertinent
  identifying information
}
\description{
Randomize Headrow Trays
}
\details{
The headrows will be returned with randomized replications of
  experimental entries occuring one after another, with check entries inserted
  at regular intervals. Note that this returns a linear arrangement of trays
  with global range values increasing to infinity (i.e., the headrow trays
  are not wrapped at the length of the experiment field). Checks are first 
  inserted at the beginning (i.e. tray 1, range 1, col 1), and then repeated 
  at the specified number of cells.
}
