% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_str.R
\name{stand_str}
\alias{stand_str}
\title{Standardize character vectors to a common format}
\usage{
stand_str(str_in, word_delim = "-", case = "upper")
}
\arguments{
\item{str_in}{Character vector}

\item{word_delim}{Character defining the word delimiter for the output character
vector, either " " (space), ".", "-", or "_"}

\item{case}{String consisting of either "upper" or "lower" to select whether
to convert output to all uppercase or all lowercase}
}
\value{
character vector of same dimensions as input vector, but with all
  elements updated to the new format
}
\description{
Standardize character vectors to a common format
}
\details{
This function somewhat crudely attempts to force character vectors
  to a common format. It does this by converting everything to uppercase, and replacing
  all common word delimiters (space, period, underscore, dash) with the selected one.
  Note that of the three, only underscores or periods will form syntactically-
  valid R variable names (override by reading in tables with check.names = FALSE). 
  However, underscores can cause problems with reading
  VCFs in PLINK, as they will by default be interpreted as delimiters for
  family and individual IDs
}
