% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonymize.R
\name{synonymize}
\alias{synonymize}
\title{Create common synonyms of strings}
\usage{
synonymize(s)
}
\arguments{
\item{s}{Character vector or string}
}
\value{
Character vector containing common synonyms of the input string or
  character vector
}
\description{
Create common synonyms of strings
}
\details{
This function takes an input string or vector of strings, 
  and creates an output vector that has uppercase, lowercase, and title case 
  versions of the input string(s) with different common delimiters 
  (space, hyphen, underscore, period). Note that the different strings present 
  in the output will have all delimiters the same (e.g. robin_sparrow_swift, 
  not robin.sparrow_swift)
}
