% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synop_dh92}
\alias{synop_dh92}
\title{Synthetic x Opata genetic map, 92 DH lines}
\format{A data frame with 7,742 rows and 4 variables:
\describe{
   \item{chr}{chromosome name}
   \item{id}{SNP identifier}
   \item{dist}{genetic position, in centiMorgans (cM)}
   \item{pos}{phyiscal position, in base pairs (bp)}
}}
\source{
\url{https://doi.org/10.1038/s41598-018-38111-3}
}
\usage{
synop_dh92
}
\description{
A dataframe adapted from Supplementary Dataset S7 of Gutierrez-Gonzalez et 
al., 2019 containing map data on SNPs generated by a PstI-MspI double digest 
genotyping-by-sequencing (GBS) protocol, with physical positions from 
the International Wheat Genome Sequencing Consortium (IWGSC) v1.0 
Chinese Spring wheat Reference Sequence and genetic positions derived from a
population of 92 doubled-haploid (DH) lines descended from the Synthetic x 
Opata cross. Gutierrez-Gonzalez et al. performed imputation using the R
package LaByRInth (https://github.com/Dordt-Statistics-Research/LaByRInth),
which is an adaptation of the LB-Impute algorithm. Several outlier SNPs were
removed from this dataset, leaving 92,272 markers.
}
\details{
This data is in the format required for PLINK .map files. However, note that
many programs that use PLINK files may require the header to be removed, and
may require genetic distances to be converted to Morgans.
}
\keyword{datasets}
