% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synop_ril173}
\alias{synop_ril173}
\title{Synthetic x Opata genetic map, 173 RIL lines}
\format{A data frame with 4,027 rows and 3 variables:
\describe{
   \item{chr}{chromosome name}
   \item{dist}{genetic position, in centiMorgans (cM)}
   \item{pos}{phyiscal position, in base pairs (bp)}
}}
\source{
\url{https://urgi.versailles.inra.fr/download/iwgsc/IWGSC_RefSeq_Annotations/v1.0/iwgsc_refseqv1.0_SynOpRIL993_GBS.zip}
}
\usage{
synop_ril173
}
\description{
A dataframe containing the chromosome, physical position, and genetic
position of SNPs generated by a PstI-MspI double digest 
genotyping-by-sequencing (GBS) protocol, comparing physical positions of SNPs
in the International Wheat Genome Sequencing Consortium (IWGSC) v1.0 
Chinese Spring wheat Reference Sequence against genetic positions in a
population of 173 recombinant inbred lines (RILs) derived from the Synthetic x Opata
cross. SNPs that mapped to different chromosomes in Chinese Spring and the
Synthetic x Opata cross have been removed, along with three outliers.
}
\keyword{datasets}
