% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2dos.R
\name{vcf2dos}
\alias{vcf2dos}
\title{Convert a VCF file to minor-allele dosage format}
\usage{
vcf2dos(vcf_dat)
}
\arguments{
\item{vcf_dat}{- A vcfR object created by the read.vcfR() function of package
vcfR}
}
\value{
dataframe in the same format as the data rows of the input VCF data, 
  except with genotype calls transformed into minor allele dosage format 
  (i.e. homozygous for minor allele = 2, het = 1, homo. for major allele = 0)
  NOTE: VCF 'FORMAT' column is not present in output, since only the GT field
  is extracted from the input VCF
}
\description{
Convert a VCF file to minor-allele dosage format
}
\details{
vcf2dos() uses vectorized operations and is therefore quite fast.
  However, since R must read an entire file into memory, large VCF files
  will cause problems, especially if the size of the file is similar to
  the amount of physical memory on the machine. In the case of large VCF 
  files, it may be best to split them apart by chromosome, or else use 
  command-line tools such as bcftools, vcftools, or PLINK 1.9
}
