% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_headrows.R
\name{randomize_headrows}
\alias{randomize_headrows}
\title{Randomize Headrow Trays}
\usage{
randomize_headrows(
  ents,
  nreps = 1,
  tray_cells = 80,
  tray_rows = 4,
  start_cell = 1,
  randomize_first = TRUE,
  checks = list("CHECK", 40),
  seed = NA
)
}
\arguments{
\item{ents}{Vector listing names of entries to include in the experiment}

\item{nreps}{Integer specifying the number of replications 
(i.e. full blocks)}

\item{tray_cells}{Integer specifying the number of cells contained by each
headrow tray}

\item{tray_rows}{Integer specifying the number of rows per range. Note that
this "rows" is shorthand for \emph{headrows}, so this might more 
intuitively be thought of as columns.}

\item{start_cell}{Integer specifying the starting number for the first cell}

\item{randomize_first}{Logical indicating whether or not to randomize the
first replication}

\item{checks}{A list with two elements:
\enumerate{
  \item A string or character vector listing the names of check(s) to
  include
  \item An integer specifying how often to repeat the checks  
}}

\item{seed}{Integer to seed the random number generator. Default NA value
will not seed the random number generator.}
}
\value{
A dataframe containing the randomized headrows and all pertinent
  identifying information
}
\description{
Randomize Headrow Trays
}
\details{
The headrows will be returned with randomized replications of
  experimental entries occuring one after another, with check entries inserted
  at regular intervals. Note that this returns a linear arrangement of trays
  with global range values increasing to infinity (i.e., the headrow trays
  are not wrapped at the length of the experiment field). Checks are first 
  inserted at the beginning (i.e. tray 1, range 1, row 1), and then repeated 
  at the specified number of cells. As noted above, the somewhat unfortunate
  nomenclature of head\emph{rows} means that what would intuitively be
  thought of as "rows" are labeled "ranges", and what would intuitively be
  thought of as "columns" are labeled "rows".
}
